requirements.txt
Pillow>=10.0.0
pycryptodome>=3.20.0
numpy>=1.26.0
google-genai>=1.0.0
requests>=2.31.0
pytest>=8.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
supabase>=2.11.0
PyJWT>=2.8.0
httpx<0.28
websockets<16
